chore: emit decl files on prerelease, combine inject version scripts
