+++ package.json
@@ -56,10 +56,10 @@
     "test:browser:watch": "web-test-runner --watch",
     "test:browser:update-snapshots": "web-test-runner --update-snapshots",
     "test:node": "mocha -r mocha-hooks.mjs './__integration__/**/*.test.js' './__tests__/**/*.test.js' './__node_tests__/**/*.test.js'",
-    "generate-docs": "node ./scripts/generateDocs.cjs",
     "serve-docs": "docsify serve docs -p 3000 -P 12345",
     "install-cli": "npm install -g $(npm pack)",
-    "release": "node scripts/inject-version.js && tsc && node ./scripts/version.cjs && npm run generate-docs && && changeset publish",
+    "prerelease": "node scripts/inject-version.js && tsc --emitDeclarationOnly",
+    "release": "npm run prerelease && changeset publish",
     "prepare": "husky install",
     "postinstall": "patch-package"
   },
