+++ tsconfig.json
@@ -7,7 +7,8 @@
     "checkJs": true,
     "strict": true,
     "noImplicitAny": true,
-    "skipLibCheck": true
+    "skipLibCheck": true,
+    "declaration": true
   },
   "include": ["types/**/*.d.ts", "lib/**/*.js"],
   "exclude": ["node_modules", "**/coverage/*"]
